+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/workshops/WorkshopDto.kt
@@ -7,7 +7,6 @@ import com.google.gson.annotations.SerializedName
  * DTO representing detailed information of a workshop received from the API.
  *
  * @param id The unique identifier of the workshop.
- * @param idTraining The unique identifier of the training.
  * @param idUser The unique identifier of the user.
  * @param image The url of the image to be upload.
  * @param name The name of the new workshop.
@@ -16,10 +15,10 @@ import com.google.gson.annotations.SerializedName
  * @param status The status of the workshop, active = 1, inactive = 0.
  * @param description The description of how is the workshop.
  * @param date The day when the workshop is going to take place.
+ * @param videoTraining The video of the training depending on the workshop
  */
 data class WorkshopDto(
     @SerializedName("idTaller") val id: String?,
-    @SerializedName("idCapacitacion") val idTraining: String?,
     @SerializedName("idUsuario") val idUser: String?,
     @SerializedName("URL") val image: String?,
     @SerializedName("nombreTaller") val name: String?,
@@ -27,5 +26,6 @@ data class WorkshopDto(
     @SerializedName("horaSalida") val finishHour: String?,
     @SerializedName("estatus") val status: Int,
     @SerializedName("Descripcion") val description: String?,
-    @SerializedName("fecha") val date: String?
+    @SerializedName("fecha") val date: String?,
+    @SerializedName("videoCapacitacion") val videoTraining: String?
 )